cactacea-core : Fix data types and original epoch time

+++ core/src/main/java/io/github/cactacea/core/domain/enums/ContentStatusType.java
@@ -3,17 +3,17 @@ package io.github.cactacea.core.domain.enums;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public enum ContentStatusType {
-    @JsonProperty("0") unchecked(0),
-    @JsonProperty("1") accepted(1),
-    @JsonProperty("2") rejected(2);
+    @JsonProperty("0") unchecked((byte)0),
+    @JsonProperty("1") accepted((byte)1),
+    @JsonProperty("2") rejected((byte)2);
 
-    private long value;
+    private byte value;
 
-    private ContentStatusType(long value) {
+    private ContentStatusType(byte value) {
         this.value = value;
     }
 
-    static public ContentStatusType forName(long value) {
+    static public ContentStatusType forName(byte value) {
         for (ContentStatusType e : values()) {
             if (e.value == value) {
                 return e;
@@ -22,7 +22,9 @@ public enum ContentStatusType {
         throw new IllegalArgumentException();
     }
 
-    public long toValue() {
+    public byte toValue() {
         return value;
     }
+
+
 }
