+++ core/src/main/java/io/github/cactacea/core/domain/enums/GroupAuthorityType.java
@@ -3,16 +3,16 @@ package io.github.cactacea.core.domain.enums;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public enum GroupAuthorityType {
-    @JsonProperty("0")owner(0),
-    @JsonProperty("1")member(1);
+    @JsonProperty("0")owner((byte)0),
+    @JsonProperty("1")member((byte)1);
 
-    private long value;
+    private byte value;
 
-    private GroupAuthorityType(long value) {
+    private GroupAuthorityType(byte value) {
         this.value = value;
     }
 
-    static public GroupAuthorityType forName(long value) {
+    static public GroupAuthorityType forName(byte value) {
         for (GroupAuthorityType e : values()) {
             if (e.value == value) {
                 return e;
@@ -21,7 +21,7 @@ public enum GroupAuthorityType {
         throw new IllegalArgumentException();
     }
 
-    public long toValue() {
+    public byte toValue() {
         return value;
     }
 }
